Data.Time.Calendar.Julian.Compat

Documentation

type Year = Integer

Year of Common Era (when positive).

type MonthOfYear = Int

Month of year, in range 1 (January) to 12 (December).

type DayOfMonth = Int

Day of month, in range 1 to 31.

type DayOfYear = Int

Day of year, in range 1 (January 1st) to 366. December 31st is 365 in a common year, 366 in a leap year.

toJulianYearAndDay :: Day -> (Year, DayOfYear)

Convert to proleptic Julian year and day format.

fromJulianYearAndDay :: Year -> DayOfYear -> Day

Convert from proleptic Julian year and day format. Invalid day numbers will be clipped to the correct range (1 to 365 or 366).

fromJulianYearAndDayValid :: Year -> DayOfYear -> Maybe Day

Convert from proleptic Julian year and day format. Invalid day numbers will return Nothing

showJulianYearAndDay :: Day -> String

Show in proleptic Julian year and day format (yyyy-ddd)

isJulianLeapYear :: Year -> Bool

Is this year a leap year according to the proleptic Julian calendar?

toJulian :: Day -> (Year, MonthOfYear, DayOfMonth)

Convert to proleptic Julian calendar.

fromJulian :: Year -> MonthOfYear -> DayOfMonth -> Day

Convert from proleptic Julian calendar. Invalid values will be clipped to the correct range, month first, then day.

pattern JulianYearMonthDay :: Year -> MonthOfYear -> DayOfMonth -> Day

Bidirectional abstract constructor for the proleptic Julian calendar. Invalid values will be clipped to the correct range, month first, then day.

fromJulianValid :: Year -> MonthOfYear -> DayOfMonth -> Maybe Day

Convert from proleptic Julian calendar. Invalid values will return Nothing.

showJulian :: Day -> String

Show in ISO 8601 format (yyyy-mm-dd)

julianMonthLength :: Year -> MonthOfYear -> DayOfMonth

The number of days in a given month according to the proleptic Julian calendar.

addJulianMonthsClip :: Integer -> Day -> Day

Add months, with days past the last day of the month clipped to the last day. For instance, 2005-01-30 + 1 month = 2005-02-28.

addJulianMonthsRollOver :: Integer -> Day -> Day

Add months, with days past the last day of the month rolling over to the next month. For instance, 2005-01-30 + 1 month = 2005-03-02.

addJulianYearsClip :: Integer -> Day -> Day

Add years, matching month and day, with Feb 29th clipped to Feb 28th if necessary. For instance, 2004-02-29 + 2 years = 2006-02-28.

addJulianYearsRollOver :: Integer -> Day -> Day

Add years, matching month and day, with Feb 29th rolled over to Mar 1st if necessary. For instance, 2004-02-29 + 2 years = 2006-03-01.

addJulianDurationClip :: CalendarDiffDays -> Day -> Day

Add months (clipped to last day), then add days

addJulianDurationRollOver :: CalendarDiffDays -> Day -> Day

Add months (rolling over to next month), then add days

diffJulianDurationClip :: Day -> Day -> CalendarDiffDays

Calendrical difference, with as many whole months as possible

diffJulianDurationRollOver :: Day -> Day -> CalendarDiffDays

Calendrical difference, with as many whole months as possible. Same as diffJulianDurationClip for positive durations.