Data.Time.Calendar.Month.Compat

Documentation

newtype Month

An absolute count of common calendar months. Number is equal to (year * 12) + (monthOfYear - 1).

Constructors

MkMonth Integer

addMonths :: Integer -> Month -> Month

diffMonths :: Month -> Month -> Integer

pattern YearMonth :: Year -> MonthOfYear -> Month

Bidirectional abstract constructor. Invalid months of year will be clipped to the correct range.

fromYearMonthValid :: Year -> MonthOfYear -> Maybe Month

pattern MonthDay :: Month -> DayOfMonth -> Day

Bidirectional abstract constructor. Invalid days of month will be clipped to the correct range.

fromMonthDayValid :: Month -> DayOfMonth -> Maybe Day

time-compat extras

fromYearMonth :: Year -> MonthOfYear -> Month

Part of YearMonth pattern

toYearMonth :: Month -> (Year, MonthOfYear)

Part of YearMonth pattern

fromMonthDay :: Month -> DayOfMonth -> Day

Part of MonthDay pattern

toMonthDay :: Day -> (Month, DayOfMonth)

Part of MonthDay pattern