Data.Time.Calendar.WeekDate.Compat

Documentation

type Year = Integer

Year of Common Era (when positive).

type WeekOfYear = Int

Week of year, by various reckonings, generally in range 0-53 depending on reckoning.

data DayOfWeek

Constructors

Monday

Tuesday

Wednesday

Thursday

Friday

Saturday

Sunday

dayOfWeek :: Day -> DayOfWeek

data FirstWeekType

Constructors

FirstWholeWeek
first week is the first whole week of the year

FirstMostWeek
first week is the first week with four days in the year

toWeekCalendar

Arguments

:: FirstWeekType
how to reckon the first week of the year

-> DayOfWeek
the first day of each week

-> Day

-> (Year, WeekOfYear, DayOfWeek)

Convert to the given kind of "week calendar". Note that the year number matches the weeks, and so is not always the same as the Gregorian year number.

fromWeekCalendar

Arguments

:: FirstWeekType
how to reckon the first week of the year

-> DayOfWeek
the first day of each week

-> Year

-> WeekOfYear

-> DayOfWeek

-> Day

Convert from the given kind of "week calendar". Invalid week and day values will be clipped to the correct range.

fromWeekCalendarValid

Arguments

:: FirstWeekType
how to reckon the first week of the year

-> DayOfWeek
the first day of each week

-> Year

-> WeekOfYear

-> DayOfWeek

-> Maybe Day

Convert from the given kind of "week calendar". Invalid week and day values will return Nothing.

ISO 8601 Week Date format

toWeekDate :: Day -> (Year, WeekOfYear, Int)

Convert to ISO 8601 Week Date format. First element of result is year, second week number (1-53), third day of week (1 for Monday to 7 for Sunday). Note that "Week" years are not quite the same as Gregorian years, as the first day of the year is always a Monday. The first week of a year is the first week to contain at least four days in the corresponding Gregorian year.

fromWeekDate :: Year -> WeekOfYear -> Int -> Day

Convert from ISO 8601 Week Date format. First argument is year, second week number (1-52 or 53), third day of week (1 for Monday to 7 for Sunday). Invalid week and day values will be clipped to the correct range.

pattern YearWeekDay :: Year -> WeekOfYear -> DayOfWeek -> Day

Bidirectional abstract constructor for ISO 8601 Week Date format. Invalid week values will be clipped to the correct range.

fromWeekDateValid :: Year -> WeekOfYear -> Int -> Maybe Day

Convert from ISO 8601 Week Date format. First argument is year, second week number (1-52 or 53), third day of week (1 for Monday to 7 for Sunday). Invalid week and day values will return Nothing.

showWeekDate :: Day -> String

Show in ISO 8601 Week Date format as yyyy-Www-d (e.g. "2006-W46-3").