Data.Time.Clock.TAI.Compat

TAI arithmetic

data AbsoluteTime

AbsoluteTime is TAI, time as measured by a clock.

taiEpoch :: AbsoluteTime

The epoch of TAI, which is 1858-11-17 00:00:00 TAI.

addAbsoluteTime :: DiffTime -> AbsoluteTime -> AbsoluteTime

addAbsoluteTime a b = a + b

diffAbsoluteTime :: AbsoluteTime -> AbsoluteTime -> DiffTime

diffAbsoluteTime a b = a - b

taiNominalDayStart :: Day -> AbsoluteTime

leap-second map type

type LeapSecondMap = Day -> Maybe Int

TAI - UTC during this day. No table is provided, as any program compiled with it would become out of date in six months.

conversion between UTC and TAI with map

utcDayLength :: LeapSecondMap -> Day -> Maybe DiffTime

utcToTAITime :: LeapSecondMap -> UTCTime -> Maybe AbsoluteTime

taiToUTCTime :: LeapSecondMap -> AbsoluteTime -> Maybe UTCTime

taiClock :: Maybe (DiffTime, IO AbsoluteTime)

TAI clock, if it exists. Note that it is unlikely to be set correctly, without due care and attention.