Data.Time.Calendar.Compat

Days

newtype Day

The Modified Julian Day is a standard count of days, with zero being the day 1858-11-17.

Constructors

ModifiedJulianDay

Fields

toModifiedJulianDay :: Integer

addDays :: Integer -> Day -> Day

diffDays :: Day -> Day -> Integer

CalendarDiffTime

data CalendarDiffDays

Constructors

CalendarDiffDays

Fields

cdMonths :: Integer

cdDays :: Integer

calendarDay :: CalendarDiffDays

calendarWeek :: CalendarDiffDays

calendarMonth :: CalendarDiffDays

calendarYear :: CalendarDiffDays

scaleCalendarDiffDays :: Integer -> CalendarDiffDays -> CalendarDiffDays

Scale by a factor. Note that scaleCalendarDiffDays (-1) will not perfectly invert a duration, due to variable month lengths.

Gregorian calendar

toGregorian :: Day -> (Year, MonthOfYear, DayOfMonth)

Convert to proleptic Gregorian calendar.

fromGregorian :: Year -> MonthOfYear -> DayOfMonth -> Day

Convert from proleptic Gregorian calendar. Invalid values will be clipped to the correct range, month first, then day.

fromGregorianValid :: Year -> MonthOfYear -> DayOfMonth -> Maybe Day

Convert from proleptic Gregorian calendar. Invalid values will return Nothing

showGregorian :: Day -> String

Show in ISO 8601 format (yyyy-mm-dd)

gregorianMonthLength :: Year -> MonthOfYear -> DayOfMonth

The number of days in a given month according to the proleptic Gregorian calendar.

addGregorianMonthsClip :: Integer -> Day -> Day

Add months, with days past the last day of the month clipped to the last day. For instance, 2005-01-30 + 1 month = 2005-02-28.

addGregorianMonthsRollOver :: Integer -> Day -> Day

Add months, with days past the last day of the month rolling over to the next month. For instance, 2005-01-30 + 1 month = 2005-03-02.

addGregorianYearsClip :: Integer -> Day -> Day

Add years, matching month and day, with Feb 29th clipped to Feb 28th if necessary. For instance, 2004-02-29 + 2 years = 2006-02-28.

addGregorianYearsRollOver :: Integer -> Day -> Day

Add years, matching month and day, with Feb 29th rolled over to Mar 1st if necessary. For instance, 2004-02-29 + 2 years = 2006-03-01.

addGregorianDurationClip :: CalendarDiffDays -> Day -> Day

Add months (clipped to last day), then add days

addGregorianDurationRollOver :: CalendarDiffDays -> Day -> Day

Add months (rolling over to next month), then add days

diffGregorianDurationClip :: Day -> Day -> CalendarDiffDays

Calendrical difference, with as many whole months as possible

diffGregorianDurationRollOver :: Day -> Day -> CalendarDiffDays

Calendrical difference, with as many whole months as possible. Same as diffGregorianDurationClip for positive durations.

isLeapYear :: Year -> Bool

Is this year a leap year according to the proleptic Gregorian calendar?

Week

data DayOfWeek

Constructors

Monday

Tuesday

Wednesday

Thursday

Friday

Saturday

Sunday

dayOfWeek :: Day -> DayOfWeek

dayOfWeekDiff :: DayOfWeek -> DayOfWeek -> Int

dayOfWeekDiff a b = a - b in range 0 to 6. The number of days from b to the next a.

firstDayOfWeekOnAfter :: DayOfWeek -> Day -> Day

The first day-of-week on or after some day

Type aliases

type DayOfMonth = Int

Day of month, in range 1 to 31.

type MonthOfYear = Int

Month of year, in range 1 (January) to 12 (December).

type Year = Integer

Year of Common Era (when positive).

pattern YearMonthDay :: Year -> MonthOfYear -> DayOfMonth -> Day

Bidirectional abstract constructor for the proleptic Gregorian calendar. Invalid values will be clipped to the correct range, month first, then day.