Data.Time.Clock.Compat

Universal Time

Time as measured by the Earth.

newtype UniversalTime

The Modified Julian Date is the day with the fraction of the day, measured from UT midnight. It's used to represent UT1, which is time as measured by the earth's rotation, adjusted for various wobbles.

Constructors

ModJulianDate

Fields

getModJulianDate :: Rational

Absolute intervals, DiffTime

data DiffTime

This is a length of time, as measured by a clock. Conversion functions will treat it as seconds. It has a precision of 10^-12 s.

secondsToDiffTime :: Integer -> DiffTime

Create a DiffTime which represents an integral number of seconds.

picosecondsToDiffTime :: Integer -> DiffTime

Create a DiffTime from a number of picoseconds.

diffTimeToPicoseconds :: DiffTime -> Integer

Get the number of picoseconds in a DiffTime.

UTCTime

data UTCTime

This is the simplest representation of UTC. It consists of the day number, and a time offset from midnight. Note that if a day has a leap second added to it, it will have 86401 seconds.

Constructors

UTCTime

Fields

utctDay :: Day

the day

utctDayTime :: DiffTime

the time from midnight, 0 <= t < 86401s (because of leap-seconds)

NominalDiffTime

data NominalDiffTime

This is a length of time, as measured by UTC. It has a precision of 10^-12 s.

Conversion functions will treat it as seconds. For example, (0.010 :: NominalDiffTime) corresponds to 10 milliseconds.

It ignores leap-seconds, so it's not necessarily a fixed amount of clock time. For instance, 23:00 UTC + 2 hours of NominalDiffTime = 01:00 UTC (+ 1 day), regardless of whether a leap-second intervened.

secondsToNominalDiffTime :: Pico -> NominalDiffTime

Create a NominalDiffTime from a number of seconds.

Since: time-1.9.1

nominalDiffTimeToSeconds :: NominalDiffTime -> Pico

Get the seconds in a NominalDiffTime.

Since: time-1.9.1

nominalDay :: NominalDiffTime

One day in NominalDiffTime.

UTC differences

addUTCTime :: NominalDiffTime -> UTCTime -> UTCTime

addUTCTime a b = a + b

diffUTCTime :: UTCTime -> UTCTime -> NominalDiffTime

diffUTCTime a b = a - b

Current time

getCurrentTime :: IO UTCTime

Get the current UTCTime from the system clock.

getTime_resolution :: DiffTime

The resolution of getSystemTime, getCurrentTime, getPOSIXTime

Type aliases

type Year = Integer

Year of Common Era.

type MonthOfYear = Int

Month of year, in range 1 (January) to 12 (December).

type DayOfMonth = Int

Day of month, in range 1 to 31.